backend/requirements.txt
Django>=5.0,<6.0
djangorestframework>=3.15
djangorestframework-simplejwt>=5.3
django-cors-headers>=4.3
django-filter>=24.0
Pillow>=10.0
python-dotenv>=1.0
google-generativeai>=0.8
groq>=0.11
numpy>=1.26
scikit-learn>=1.4
gunicorn>=22.0
whitenoise>=6.7
pdfplumber>=0.11
PyMuPDF>=1.25
chromadb>=0.5
tiktoken>=0.9
